item cards moved to common
