+++ src/components/common/itemCard/mainSections/ManagementMainSection.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import SpecificationCard from './common/SpecificationCard';
-import { colors, specificationsTypes } from '../../../config/constants';
-import Title from '../../generic/Title';
+import { colors, specificationsTypes } from '../../../../config/constants';
+import Title from '../../../generic/Title';
 
 const Container = styled.div`
  	display: flex;
