+++ src/components/common/itemCard/priceSections/CommonPriceSection.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { colors } from '../../../config/constants';
+import PropTypes from 'prop-types';
+import { colors } from '../../../../config/constants';
 
 const Container = styled.div`  
 	display: flex;
@@ -16,23 +17,30 @@ const SubPriceText = styled.span`
 	text-align: center;
 	color: ${colors.textLightGray};
 	font-size: 14px;
-	margin-top: 15px;`;
+	margin-top: 30px;`;
 
-const UpPriceText = styled.span`  
-	text-align: center;
-	color: ${colors.textLightGray};
-	font-size: 14px;`;
-
-const CloudGuardPriceSection = ({ price, additionalText }) => {
+const CommonPriceSection = ({ price, additionalText, ...otherProps }) => {
 
 	return (
 		<Container>
-			<UpPriceText>Starting at</UpPriceText>
 			<PriceText>${ price }</PriceText>
 			<SubPriceText>{ additionalText }</SubPriceText>
 		</Container>
 	);
 };
 
-export default CloudGuardPriceSection;
+CommonPriceSection.defaultProps = {
+	price: 70,
+	additionalText: '1 PC/YEAR & 1 Mobile/Year',
+};
+
+CommonPriceSection.propTypes = {
+	price: PropTypes.oneOfType([
+		PropTypes.string,
+		PropTypes.number
+	]),
+	additionalText: PropTypes.string,
+};
+
+export default CommonPriceSection;
